feat: implement full domain and application layers for car-service
Added complete domain model including Car aggregate, value objects (CarId, Price, CarCategory, AvailabilityStatus) and domain events.

Introduced domain ports for persistence, querying, event publishing and reservation policy integration.

Implemented full application layer with CQRS: command handlers in CarApplicationService and query handlers in CarQueryService, including transactional boundaries and traceId/spanId logging.

Added validation-ready command/query DTOs and separated write/read use cases.

Implemented Saga process manager (ReservationSagaHandler) coordinating reservation and payment events with car availability logic.

+++ car-service/src/main/java/car_rental_app/domain/model/Car.java
@@ -21,7 +21,7 @@ public class Car {
         this.category = category;
         this.price = price;
         this.status = AvailabilityStatus.AVAILABLE;
-        domainEvents.add(new CarCreatedEvent(id.value(), category.name(), BigDecimal.valueOf(price.value())));
+        domainEvents.add(new CarCreatedEvent(id.value(), category.name(), price.value()));
     }
 
     public void markAsUnavailable() {
@@ -38,7 +38,7 @@ public class Car {
 
     public void changePrice(Price newPrice) {
         this.price = newPrice;
-        domainEvents.add(new CarPriceChangedEvent(id.value(), BigDecimal.valueOf(newPrice.value())));
+        domainEvents.add(new CarPriceChangedEvent(id.value(), newPrice.value()));
     }
 
     public List<Object> getDomainEvents() {
